fixed beta test bug

+++ core/database/migrations/2014_09_28_122406_create_prodis_table.php
@@ -18,8 +18,8 @@ class CreateProdisTable extends Migration
             $table->string('nama');
             $table->Integer('fakultas_id')->nullable();
             $table->tinyInteger('status');
-            $table->integer('created_by')->nullable();
-            $table->integer('updated_by')->nullable();
+            $table->bigInteger('created_by')->nullable();
+            $table->bigInteger('updated_by')->nullable();
             $table->timestamps();
         });
     }
